+++ common/src/main/java/net/blay09/mods/hardcorerevival/HardcoreRevival.java
@@ -2,9 +2,11 @@ package net.blay09.mods.hardcorerevival;
 
 import net.blay09.mods.balm.Balm;
 import net.blay09.mods.balm.core.BalmRegistrars;
+import net.blay09.mods.hardcorerevival.command.KnockoutCommand;
 import net.blay09.mods.hardcorerevival.command.ReviveCommand;
 import net.blay09.mods.hardcorerevival.compat.Compat;
 import net.blay09.mods.hardcorerevival.config.HardcoreRevivalConfig;
+import net.blay09.mods.hardcorerevival.config.HardcoreRevivalRules;
 import net.blay09.mods.hardcorerevival.handler.*;
 import net.blay09.mods.hardcorerevival.network.ModNetworking;
 import net.blay09.mods.hardcorerevival.stats.ModStats;
@@ -19,10 +21,12 @@ public class HardcoreRevival {
 
     public static void initialize(BalmRegistrars registrars) {
         HardcoreRevivalConfig.initialize();
+        HardcoreRevivalRules.initialize();
 
         ModNetworking.initialize(Balm.networking());
         registrars.customStats(ModStats::initialize);
 
+        Balm.commands().register(KnockoutCommand::register);
         Balm.commands().register(ReviveCommand::register);
 
         KnockoutHandler.initialize();